Create scanpipeline.sh
This script integrates scancode results with OpossumUI by using the Opossum-coversion-tool opossum-lib-sh.exe. It is meant to be used by people who are not using CLI tools everyday and thereby adheres to a simple question-answer-format. However, it requires scancode-toolkit, OpossumUI and Opossum.lib.hs to be installed, preferably in one (parent) directory. The script is prepared to offer ORT integration, but since ORT was very buggy in testing it is not recommended to use, however, if a solid environment is set up, it might come in handy.